backend/requirements.txt
fastapi
uvicorn
python-multipart
tensorflow
numpy
pillow
pandas